added park data service

+++ GoToParks.Mobile/GoToParks.Core/Model/Park.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string Address { get; set; }
-        public decimal Lat { get; set; }
-        public decimal Long { get; set; }
+        public double Lat { get; set; }
+        public double Long { get; set; }
     }
 }
